+++ tests/tests_fabric/strategies/test_strategy.py
@@ -18,7 +18,10 @@ import pytest
 import torch
 from torch import nn
 
+from lightning.fabric.plugins import DoublePrecision, HalfPrecision, Precision
 from lightning.fabric.strategies import SingleDeviceStrategy
+from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_0
+from tests_fabric.helpers.runif import RunIf
 
 
 @pytest.mark.parametrize("is_rank_zero", [True, False])
@@ -120,3 +123,82 @@ def test_load_checkpoint_strict_loading(tmp_path):
     strategy.checkpoint_io.load_checkpoint = load_checkpoint_mock
     with pytest.raises(KeyError, match="contains a key 'c' that does not exist"):
         strategy.load_checkpoint(tmp_path, requested_state)
+
+
+@RunIf(min_torch="1.13")
+@pytest.mark.parametrize(
+    "device",
+    [
+        "cpu",
+        pytest.param("cuda:0", marks=RunIf(min_cuda_gpus=1)),
+        pytest.param("mps:0", marks=RunIf(mps=True)),
+    ],
+)
+@pytest.mark.parametrize(
+    ("precision", "dtype"),
+    [
+        (Precision(), torch.float32),
+        (HalfPrecision("16-true"), torch.float16),
+        pytest.param(HalfPrecision("bf16-true"), torch.bfloat16, marks=RunIf(mps=False)),
+        pytest.param(DoublePrecision(), torch.float64, marks=RunIf(mps=False)),
+    ],
+)
+@pytest.mark.parametrize("empty_weights", [None, True, False])
+def test_module_init_context(device, precision, dtype, empty_weights, monkeypatch):
+    """Test that the module under the init-module-context gets moved to the right device and dtype."""
+    init_mock = Mock()
+    monkeypatch.setattr(torch.Tensor, "uniform_", init_mock)
+
+    device = torch.device(device)
+    strategy = SingleDeviceStrategy(device=device, precision=precision)  # surrogate class to test base class
+    with strategy.module_init_context(empty_weights=empty_weights):
+        module = torch.nn.Linear(2, 2)
+
+    expected_device = device if _TORCH_GREATER_EQUAL_2_0 else torch.device("cpu")
+    assert module.weight.device == module.bias.device == expected_device
+    assert module.weight.dtype == module.bias.dtype == dtype
+    if not empty_weights:
+        init_mock.assert_called()
+    else:
+        init_mock.assert_not_called()
+
+
+@mock.patch("lightning.fabric.strategies.strategy._TORCH_GREATER_EQUAL_1_13", False)
+def test_module_init_context_empty_weights_support():
+    strategy = SingleDeviceStrategy()  # surrogate class to test base class
+    with pytest.raises(NotImplementedError, match="`empty_weights=True` requires PyTorch >= 1.13"):
+        with strategy.module_init_context(empty_weights=True):
+            pass
+
+
+@pytest.mark.parametrize(
+    "device",
+    [
+        "cpu",
+        pytest.param("cuda:0", marks=RunIf(min_cuda_gpus=1)),
+        pytest.param("mps:0", marks=RunIf(mps=True)),
+    ],
+)
+@pytest.mark.parametrize(
+    ("precision", "dtype"),
+    [
+        (Precision(), torch.float32),
+        (HalfPrecision("16-true"), torch.float16),
+        pytest.param(HalfPrecision("bf16-true"), torch.bfloat16, marks=RunIf(mps=False)),
+        pytest.param(DoublePrecision(), torch.float64, marks=RunIf(mps=False)),
+    ],
+)
+def test_tensor_init_context(device, precision, dtype):
+    """Test that tensors under the init-tensor-context get moved to the right device and dtype."""
+    device = torch.device(device)
+    strategy = SingleDeviceStrategy(device=device, precision=precision)  # surrogate class to test base class
+    with strategy.tensor_init_context():
+        tensor0 = torch.tensor(42.0)
+        tensor1 = torch.tensor(42)
+        tensor2 = torch.tensor(42.0, dtype=torch.half)
+
+    expected_device = device if _TORCH_GREATER_EQUAL_2_0 else torch.device("cpu")
+    assert tensor0.device == tensor1.device == tensor2.device == expected_device
+    assert tensor0.dtype == dtype
+    assert tensor1.dtype == torch.long  # `.init_tensor()` only affects floating point dtypes
+    assert tensor2.dtype == torch.half  # this tensor was created with an explicit dtype assignment
